Multithreading is now done by UMAP itself.
Feel free to share any ideas how UMAP can be further improved.