@tjjordan 2023-01-14
Here are some benchmarks.
They were done on the VM, so they are not reproducible.

This branch:
```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                    Time             CPU   Iterations
------------------------------------------------------------------------------------------------------------------------------
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/1/real_time  4086306461 ns    497560512 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/2/real_time  4878144534 ns    475130036 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/4/real_time  2669281976 ns    484542436 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/8/real_time  3954180678 ns    453432220 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/16/real_time 3157435189 ns    473646832 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/32/real_time 3584578933 ns    468059560 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/1/real_time      77163689735 ns   1696047765 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/2/real_time      121017392113 ns   1713934639 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/4/real_time      56637528519 ns   1642359962 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/8/real_time      114150112228 ns   1763865333 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/16/real_time     30741021320 ns   1604396735 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/32/real_time     70098785277 ns   1618349068 ns            1
```

Master branch:
```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                    Time             CPU   Iterations
------------------------------------------------------------------------------------------------------------------------------
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/1/real_time  3162340133 ns    494435726 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/2/real_time  2858898059 ns     35158599 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/4/real_time  1052242220 ns     18879209 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/8/real_time   583029359 ns     11660270 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/16/real_time  519512107 ns     13320299 ns            2
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_SEQUENTIAL/100/32/real_time  441035322 ns     14606778 ns            2
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/1/real_time      105752606488 ns   1739457155 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/2/real_time      16302645359 ns     99340415 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/4/real_time      21559800354 ns    144750315 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/8/real_time      11245284588 ns    283464748 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/16/real_time     6620755381 ns    546125001 ns            1
FileIOMicroReadBenchmarkFixture/UMAP_ATOMIC_MAP_PRIVATE_RANDOM/100/32/real_time     5592984798 ns    793246578 ns            1
```