@tjjordan 2023-01-14
The old UMap was better..
But to be sure we need results from Koroneia.